Navbar Button fixes/New Tabs
I wanted to fix the buttons on the top navBar, the method I used before was pretty sloppy and not a good way to have skewed overlapping buttons, so this is more of a proper fixed and makes it 100x easier to move them, and more importantly add more. I also fixed the cube on the mobile site (I think?) scaling down to keep it from clipping the menu, and then future proof for when I add homescreen cards and other additions.