UI uverhaul
Over the course of this project I've moved css elements all over the place, and I will certainly continue to do so. However, I managed to remove and condense a lot of css properties and either replace them with javascript, or just manage to trash ones that have no reason of staying. Currently unaware of any issues and will continue to watch the font issues ive been having. I need to find a better way for browsers to load the text style faster, or make a failsafe that makes the page buffer for a few seconds before displaying literally anything.